MAKE `AssertInterface` trait with custom assertion methods
