+++ src/Utils/Traits/InterfaceTest.php
@@ -6,27 +6,7 @@ trait InterfaceTest
 {
     // todo: add tests
 
-    /** @test */
-    public function interface_exist()
-    {
-        $this->assertTrue(interface_exists($this->interface()));
-    }
-
-    /** @test */
-    public function interface_is_implemented()
-    {
-        foreach ($this->classes() as $class) {
-            $this->assertArrayHasKey($this->interface(), class_implements($class));
-        }
-    }
-
-    /** @test */
-    public function interface_methods_exist()
-    {
-        foreach ($this->methods() as $method) {
-            $this->assertTrue(method_exists($this->interface(), $method));
-        }
-    }
+    use AssertInterface;
 
     /**
      * Retrieve the interface class that should be tested.
